% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta13.r
\name{convolutionSum}
\alias{convolutionSum}
\title{Sum of two random discrete variables using convolution formula}
\usage{
convolutionSum(f, g)
}
\arguments{
\item{f}{PDF of variable X.}

\item{g}{PDF of variable Y.}
}
\value{
PDF of the variable Z = X + Y.
}
\description{
This function calculates the sum of two random discrete variables X and Y, returning the PDF of Z = X + Y.
}
\examples{
x <- function(t) dnorm(t, 1, 0.5)
y <- function(t) dlnorm(t, 1.5, 0.75)
z <- convolutionSum(x, y)
z <- Vectorize(z)
}
\keyword{convolution}
\keyword{discrete}
\keyword{random}
\keyword{sum}
\keyword{variable}
