% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta4.r
\name{integratePDF}
\alias{integratePDF}
\title{Integrate PDF}
\usage{
integratePDF(X)
}
\arguments{
\item{X}{The point in which the integral will be calculated.}
}
\value{
The value of the integral calculated.
}
\description{
This functions integrates the PDF. It can be used to get the CDF.
}
\keyword{cdf}
\keyword{plot}
\keyword{plotcdf}
