% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta4.r
\name{inverse}
\alias{inverse}
\title{Inverse}
\usage{
inverse(f, u, lower = -100, upper = 100)
}
\arguments{
\item{lower}{The lower part of the codomain.}

\item{upper}{The upper part of the codomain.}

\item{f(x)}{A function.}

\item{x}{The point in which the inverse function will be calculated (will raise error if the inverse cannot be calculated in that point).}
}
\value{
Dataframe containing the root of f(x) - u.
}
\description{
This function calculates the inverse of a function in a certain point.
}
\keyword{function}
\keyword{inv}
\keyword{inverse}
