% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta4.r
\name{plotCDF}
\alias{plotCDF}
\title{Plot CDF}
\usage{
plotCDF(x_min, x_max)
}
\arguments{
\item{x_min}{The minimum point of the range over which the function will be plotted.}

\item{x_max}{The maximum point of the range over which the function will be plotted.}
}
\description{
This functions plots the CDF for a random discrete variable.
}
\keyword{cdf}
\keyword{plot}
\keyword{plotcdf}
