% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta9.r
\name{randomVariableGenerator}
\alias{randomVariableGenerator}
\title{Random Variable Generation Function}
\usage{
randomVariableGenerator(f, n, x_min, x_max)
}
\arguments{
\item{f}{The PDF used to generate n variables.}

\item{n}{The number of the variables to be generated.}

\item{x_min, x_max}{The PDF's domain range.}
}
\value{
An array with n random generated values.
}
\description{
This function calculates a set of n random variables using the inverse method.
}
\examples{
f <- function(x) (1/2 * sin (x)) * (0 < x & x <= pi)
randomVariableGenerator (f, 100, 0, pi)
}
\keyword{generation}
\keyword{pdf}
\keyword{random}
\keyword{variable}
