% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta4.r
\name{integratePDF}
\alias{integratePDF}
\title{Integrate PDF}
\usage{
integratePDF(f, X)
}
\arguments{
\item{f}{The PDF for a continuous random variable.}

\item{X}{The point in which the integral will be calculated.}
}
\value{
The value of the integral calculated.
}
\description{
This functions integrates the PDF of a certain continuous random variable. It can be used to calculate the CDF for that variable in a certain point.
}
\keyword{cumulative}
\keyword{density}
\keyword{distribution}
\keyword{function}
\keyword{integrate}
\keyword{pdf}
\keyword{probability}
