% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta4.r
\name{plotCDF}
\alias{plotCDF}
\title{Plot CDF}
\usage{
plotCDF(f, x_min, x_max)
}
\arguments{
\item{f}{The PDF for a continuous random variable.}

\item{x_min, x_max}{The range over which the CDF will be plotted.}
}
\description{
This functions plots the CDF for a random discrete variable.
}
\examples{
f <- function(x) (1/2 * sin (x)) * (0 < x & x <= pi)
plotCDF(f, 0, pi)
}
\keyword{cdf}
\keyword{cumulative}
\keyword{distribution}
\keyword{function}
\keyword{plot}
\keyword{plotcdf}
