% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta4.r
\name{plotPDF}
\alias{plotPDF}
\title{Plot PDF}
\usage{
plotPDF(x_min, x_max)
}
\arguments{
\item{x_min}{The lower part of the PDF's domain.}

\item{x_max}{The upper part of the PDF's domain.}
}
\description{
This functions plots the PDF for a random discrete variable.
}
\keyword{pdf}
\keyword{plot}
\keyword{plotpdf}
