% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta9.r
\name{randomVariableGenerator}
\alias{randomVariableGenerator}
\title{Random Variable Generation Function}
\usage{
randomVariableGenerator(n, x_min, x_max)
}
\arguments{
\item{n}{The number of the variables to be generated.}

\item{x_min}{The lower part of the domain of the PDF of the variable.}

\item{lower}{The upper part of the domain of the PDF of the variable.}
}
\value{
An array containin n random variables.
}
\description{
This function calculates a set of n random variables using the inverse method.
}
\keyword{generation}
\keyword{pdf}
\keyword{random}
\keyword{variable}
